refactor : sysout제거 불필요한 함수 제거
sysout을 제거하고 불필요한 함수를 제거하였습니다.
