+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/controller/CommentController.java
@@ -33,5 +33,4 @@ public class CommentController {
         commentService.deleteComment(commentId);
         return Result.success();
     }
-
 }
