+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/dto/response/ResponseCreateComment.java
@@ -1,5 +1,6 @@
 package com.example.sosikcommunityservice.dto.response;
 
+import com.example.sosikcommunityservice.model.entity.CommentEntity;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
@@ -13,4 +14,14 @@ public record ResponseCreateComment(
         String content,
         LocalDateTime createdAt
 ) {
+    public static ResponseCreateComment responseCreate(CommentEntity commentEntity, ResponseGetMember responseGetMember) {
+        return ResponseCreateComment.builder()
+                .id(commentEntity.getId())
+                .memberId(commentEntity.getMemberId())
+                .nickname(responseGetMember.nickname())
+                .profileImage(responseGetMember.profileImage())
+                .content(commentEntity.getContent())
+                .createdAt(commentEntity.getCreatedAt())
+                .build();
+    }
 }
