+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/dto/response/ResponseGetPost.java
@@ -1,10 +1,13 @@
 package com.example.sosikcommunityservice.dto.response;
 
-import com.example.sosikcommunityservice.model.entity.CommentEntity;
+import com.example.sosikcommunityservice.model.entity.PostEntity;
 import lombok.Builder;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Builder
 public record ResponseGetPost(
@@ -18,4 +21,40 @@ public record ResponseGetPost(
         List<ResponseGetComment> comments
 
    ) {
+
+   public static ResponseGetPost responseGetPost(PostEntity postEntity, ResponseGetMember responseGetMember) {
+      return new ResponseGetPost(
+              postEntity.getMemberId(),
+              responseGetMember.nickname(),
+              responseGetMember.profileImage(),
+              postEntity.getTitle(),
+              postEntity.getContent(),
+              postEntity.increase(postEntity.getHits()),
+              postEntity.getCreatedAt(),
+              postEntity.getComments().stream()
+                      .map(commentEntity -> {
+                         String finalUrl = UriComponentsBuilder.fromHttpUrl("http://localhost:9000/members/v1/"+commentEntity.getMemberId())
+                                 .build()
+                                 .toUriString();
+                         WebClient webClient = WebClient.create();
+                         // GET 요청 보내기
+                         ResponseGetMember leaveMember = webClient.get()
+                                 .uri(finalUrl)
+                                 .retrieve()
+                                 .bodyToMono(ResponseGetMember.class)
+                                 .block();
+                         return ResponseGetComment.builder()
+                                 .id(commentEntity.getId())
+                                 .communityId(postEntity.getId())
+                                 .nickname(leaveMember.nickname())
+                                 .profileImage(leaveMember.profileImage())
+                                 .memberId(commentEntity.getMemberId())
+                                 .content(commentEntity.getContent())
+                                 .createdAt(commentEntity.getCreatedAt())
+                                 .build();
+                      })
+                      .collect(Collectors.toList())
+      );
+   }
+
 }
