+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/dto/response/Result.java
@@ -21,18 +21,4 @@ public class Result<T>{
         return new Result<T>("데이터 전송에 성공하였습니다!",result);
     }
 
-    public String toStream(){
-        if(result ==null){
-            return "{" +
-                    "\"resultCode\":" + "\""+resultCode + "\"," +
-                    "\"result\":" + null + "}";
-        }
-        {
-            return "{" +
-                    "\" resultCode\":" +"\""+resultCode + "\"," +
-                    "\"result\" :" + "\"" + result + "\"" + "}";
-
-        }
-
-    }
 }
