+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/exception/ErrorCode.java
@@ -7,7 +7,6 @@ import org.springframework.http.HttpStatus;
 @Getter
 @AllArgsConstructor
 public enum ErrorCode {
-
     POST_NOT_FOUND(HttpStatus.NOT_FOUND, "게시글을 찾지 못했습니다."),
     COMMENT_NOT_FOUND(HttpStatus.NOT_FOUND, "댓글을 찾지 못했습니다."),
     UNAUTHORIZED_UPDATE(HttpStatus.NOT_FOUND, "권한이 없어 수정할 수 없습니다.");
