+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/model/entity/PostEntity.java
@@ -1,10 +1,10 @@
 package com.example.sosikcommunityservice.model.entity;
 
+import com.example.sosikcommunityservice.dto.request.RequestCreatePost;
 import com.example.sosikcommunityservice.dto.request.RequestUpdatePost;
 import com.example.sosikcommunityservice.dto.response.*;
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.ReadOnlyProperty;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -28,7 +28,7 @@ import java.util.stream.Collectors;
                 )
         }
 )
-// 작성자,내용,제목
+
 public class PostEntity extends AuditingFields {
 
     @Id
@@ -72,43 +72,18 @@ public class PostEntity extends AuditingFields {
         this.content = requestUpdatePost.content();
     }
 
-    public Integer plushit(Integer hits) {
+    public Integer increase(Integer hits) {
         this.hits = hits + 1;
         return hits;
     }
 
-    public static ResponseGetPost responseGetPost(PostEntity postEntity, ResponseGetMember responseGetMember) {
-        return new ResponseGetPost(
-                postEntity.getMemberId(),
-                responseGetMember.nickname(),
-                responseGetMember.profileImage(),
-                postEntity.getTitle(),
-                postEntity.getContent(),
-                postEntity.plushit(postEntity.getHits()),
-                postEntity.getCreatedAt(),
-                postEntity.getComments().stream()
-                        .map(commentEntity -> {
-                            String finalUrl = UriComponentsBuilder.fromHttpUrl("http://localhost:9000/members/v1/"+commentEntity.getMemberId())
-                                    .build()
-                                    .toUriString();
-                            WebClient webClient = WebClient.create();
-                            // GET 요청 보내기
-                            ResponseGetMember leaveMember = webClient.get()
-                                    .uri(finalUrl)
-                                    .retrieve()
-                                    .bodyToMono(ResponseGetMember.class)
-                                    .block();
-                            return ResponseGetComment.builder()
-                                    .id(commentEntity.getId())
-                                    .communityId(postEntity.getId())
-                                    .nickname(leaveMember.nickname())
-                                    .profileImage(leaveMember.profileImage())
-                                    .memberId(commentEntity.getMemberId())
-                                    .content(commentEntity.getContent())
-                                    .createdAt(commentEntity.getCreatedAt())
-                                    .build();
-                        })
-                        .collect(Collectors.toList())
-        );
+    public static PostEntity create(Long memberId, RequestCreatePost postDTO) {
+        return PostEntity.builder()
+                .memberId(memberId)
+                .title(postDTO.title())
+                .content(postDTO.content())
+                .hits(0)
+                .build();
     }
+
 }
