+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/repository/CommentRepository.java
@@ -6,5 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
-    List<CommentEntity> findByCommunityId(Long communityId);
 }
