+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/repository/PostRepository.java
@@ -1,7 +1,6 @@
 package com.example.sosikcommunityservice.repository;
 
 import com.example.sosikcommunityservice.model.entity.PostEntity;
-import com.example.sosikcommunityservice.model.entity.CommentEntity;
 import com.example.sosikcommunityservice.repository.querycustom.PostRepositoryCustom;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -10,6 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface PostRepository extends JpaRepository<PostEntity, Long>, PostRepositoryCustom {
-    List<PostEntity> findAllByOrderByCreatedAtDesc();
     Slice<PostEntity> findAllByOrderByIdDesc(Pageable pageable);
 }
