+++ sosik-community-service/src/main/java/com/example/sosikcommunityservice/repository/querycustom/PostRepositoryCustom.java
@@ -1,9 +1,6 @@
 package com.example.sosikcommunityservice.repository.querycustom;
 
-import com.example.sosikcommunityservice.dto.request.GetPostSliceCondition;
-import com.example.sosikcommunityservice.dto.response.ResponseGetPostList;
 import com.example.sosikcommunityservice.model.entity.PostEntity;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
